Improvements to HostEd (#2768)
This PR proposes some changes to the Hosted classes

**Update simpleRPC to master**

With clang-tidy #2648 some failures occur in the `simpleRPC` library. Updating this to current master solves the issues.
I've also put all the code into the `simpleRPC` namespace which sorts out the conflict with `Vector` but makes it available for use if required.

**Tidy HostTests module**

Also decode the anonymous 'packet' blob so we can see what's in it and compare with spec. should we wish to update it.

**Use abstract base class for callbacks**

Simplifies implementation since usually all callbacks are required.
Compiler ensures all methods have implementations.
`clang-tidy` gave potential memory leak indication because of Delegates. Not sure if that's real but this fixes it.