Add shell functions to check for installed tools
The new shell function `check_for_installed_tools` can be used to
check for installed tools. the function takes a list of required
executables for argument. Shell process is aborted if any of the tools
is missing. All tools missing in the list are reported.

Similar the the shell function `check_for_installed_files` checks for
installed files, here used to make sure header files are installed.