Add sample to demonstrate basic dual-core operation (#2949)
This PR adds a sample application to show how to run code on the second CPU of an ESP32, RP2040 and RP2350 devices.

Issue #2943 highlights the sort of problem which occurs with the ESP32 when attempting to use timers to run code very frequently, where the CPU time required to service an interrupt and get the code to run is too long. The problem is the overhead impose by FreeRTOS and the nature of multitasking.

A solution to this is to run time-sensitive code on the second CPU. Polled timers are a good choice here as they don't use interrupts at all.