Fix exception due to static initialiser in ISR
Non-trivial static variables may require an initialisation routine
which is invoked on first call. This happens in the `hwTimerCallback`
routine when the `ElapseTimer` constructor is called, but that is unsafe
to do from interrupt context.

Moving this to a global variable fixes the issue.

The exception may vary, but I'm getting:

```
assert failed: xQueueSemaphoreTake queue.c:1718 (!( ( xTaskGetSchedulerState() == ( ( BaseType_t ) 0 ) ) && ( xTicksToWait != 0 ) ))
```